pyproject.toml
[project]
name = "tts"
version = "0.1.0"
description = "Convert text or Markdown to speech using ElevenLabs"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs",
    "python-dotenv",
]

[project.scripts]
tts = "tts:main"
